+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Since v0.0.10, this Changelog is formatted according to the [Common Changelog][c
 ## UNRELEASED
 
 ### Fixed
+- ([#1194](https://github.com/demos-europe/demosplan-ui/pull/1194)) DpTreeListNode: Expand the DpTreeListNode only when children exist ([@sakutademos](https://github.com/sakutademos)
 - ([#1192](https://github.com/demos-europe/demosplan-ui/pull/1192)) DpTimePicker: Validate manual input with available hours and minutes ([@gruenbergerdemos](https://github.com/gruenbergerdemos))
 
 ## v0.3.48 - 2025-02-20